Refactor test fixtures: add db_session fixture and remove integration conftest.py
