--- tests/integration/conftest.py
@@ -1,15 +1,0 @@
-"Shared across all infrastructure layer tests."
-import pytest
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from backend.app.infrastructure.orm_models import Base
-
-@pytest.fixture(scope="function")
-def db_session():
-    engine = create_engine("sqlite:///:memory:")
-    Base.metadata.create_all(engine)
-    Session = sessionmaker(bind=engine)
-    session = Session()
-    yield session
-    session.close()
-    Base.metadata.drop_all(engine)
